Add Arch CFN parameter required for ARM support
This commit adds the Arch CFN parameter to the configuration code.
This new parameter is a derived entirely from other parameters in the
configuration file, and is thus not permitted to be configured there.

Also included are validators that cause cluster creation to fail fast
when the architecture-specific components are not compatible with each
other or with the features enabled by the configuration file. The
specific validators are:
* Ensure the master and compute instance types have compatible instance
  architectures
* Ensure a custom AMI supports the architecture implied by the instance
  types
* Ensure that ARM instances are used only on OSs for which ARM base
  AMIs are available
* Ensure that incompatible features like disabling hyperthreading, FSx,
  or the Intel HPC platform are not enabled when using ARM instance
  types

This commit also adds the Arch parameter to the CFN template and
modifies AMI selection such that a architecture-specific mapping is
used depending on the value of this new parameter.

Signed-off-by: Tim Lane <tilne@amazon.com>