Account for ARM mapping when generating AMI lists
Note that placeholder AMI IDs are used for unsupported ARM OSs.

This is necessary because of the way CloudFormation evaluates the
template. A conditional expression chooses an AMI from either the
x86_64 or ARM mappings. CloudFormation evaluates both of these
expressions before choosing one. If one of the mappings doesn't
contain a result (as would be expected for OSes for which we don't
build ARM AMIs if placeholders are not used) then cluster creation
fails.

Signed-off-by: Tim Lane <tilne@amazon.com>